Crate teensy4_bsp

A Rust board support package (BSP) for the Teensy 4.

The BSP is mainly a pass-through of the imxrt-hal hardware abstraction layer. The BSP restricts the processor pads that are available, since the physical Teensy only has a few user-accessible pins. From these pins, you may construct peripherals and perform I/O.

The BSP also exposes a USB logging interface. See the usb module for more details.

The BSP does assume some facilities of the processor, both which are required for the USB stack. Each are controllable through feature-flags. Each feature is on by default.

it registers the SysTick exception handler, and configures SYSTICK for a 1ms interrupt. Enabled with the "systick" feature, which is on by default.

it registers the USB_OTG1 interrupt, and uses the USB1 peripheral for logging. Enabled with the "usb-logging" feature, which is on by default. Depends on the "systick" feature.

These peripherals and capabilities are not exported from the BSP. If a user also registers a SysTick or USB_OTG1 handler, it may result in a duplicate definition error.

Re-exports

The BSP re-exports the following:

the teensy4-rt crate, as rt

the imxrt-hal crate, as hal

See the accompanying documentation of each crate for more information.

For simplicity, there may be other choice APIs from either crate that are re-exported in the BSP namespace.

Although it's not exported publicly, the BSP crate links in the teensy4-fcb crate, which provides a Firmware Configuration Block (FCB) specific to the Teensy 4. See the teensy4-fcb crate for details on FCBs.

Physical Pins to Pads and Alternative Functions

The sparse table below describes the Teensy 4 pins, the pad ID, and some of the notable alternative functionalities for each pin. We add entries to the table as we add capabilities to the underlying HAL crate. Contributions to complete this table are welcome! If a pad's alternatives are not listed here, consult the iMXRT1060 reference manual.

Pin	Pad ID	Alt0	Alt1	Alt2	Alt3	Alt4	Alt5	Alt6	Alt7	Alt8	Alt9

0	AD_B0_03			UART6_RX		FlexPWM1_1_X

1	AD_B0_02			UART6_TX		FlexPWM1_0_X

2	EMC_04		FlexPWM4_2_A

3	EMC_05		FlexPWM4_2_B

4	EMC_06		FlexPWM2_0_A

5	EMC_08		FlexPWM2_1_A

6	B0_10			FlexPWM2_2_A

7	B1_01			UART4_RX				FlexPWM1_3_B

8	B1_00			UART4_TX				FlexPWM1_3_A

9	B0_11			FlexPWM2_2_B

10	B0_00				SPI4_PCS0

11	B0_02				SPI4_SDO

12	B0_01				SPI4_SDI

13	B0_03				SPI4_SCK		GPIO2_3 (LED)

14	AD_B1_02			UART2_TX

15	AD_B1_03			UART2_RX

16	AD_B1_07		I2C3_SCL	UART3_RX

17	AD_B1_06		I2C3_SDA	UART3_TX

18	AD_B1_01				I2C1_SDA

19	AD_B1_00			UART2_CTS	I2C1_SCL

20	AD_B1_10			UART8_TX

21	AD_B1_11			UART8_RX

22	AD_B1_08		FlexPWM4_0_A

23	AD_B1_09		FlexPWM4_1_A

24	AD_B0_12	I2C4_SCL		UART1_TX		FlexPWM1_2_X

25	AD_B0_13	I2C4_SDA		UART1_RX		FlexPWM1_3_X

26	AD_B1_14

27	AD_B1_15

28	EMC_32		FlexPWM3_1_B	UART7_RX

29	EMC_31		FlexPWM3_1_A	UART7_TX	SPI1_PCS1

30	EMC_37

31	EMC_36

32	B0_12

33	EMC_07		FlexPWM2_0_B

34	SD_B0_03		FlexPWM1_1_B			SPI1_SDI

35	SD_B0_02		FlexPWM1_1_A			SPI1_SDO

36	SD_B0_01		FlexPWM1_0_B	I2C3_SDA		SPI1_PCS0

37	SD_B0_00		FlexPWM1_0_A	I2C3_SCL		SPI1_SCK

38	SD_B0_05		FlexPWM1_2_B	UART8_RX

39	SD_B0_04		FlexPWM1_2_A	UART8_TX

References:

Teensy 4.0 Schematic Diagram

Examples

See the teensy4-examples crate for build-able, run-able examples. The examples utilize this BSP crate to blink LEDs, establish timers, and log data over USB.

Notice of alpha status

We've made some assumptions in this MVP BSP.

SYSTICK and delay implementation is very naive. Do not run for 49 continuous days, or risk a millisecond counter wrap-around.

Re-exports

pub use cortex_m_rt as rt;

pub use imxrt_hal as hal;

Modules

usb
Teensy 4 USB, taken from the original Teensy 4 C libraries.

Structs

Peripherals
All peripherals available on the Teensy4

Pins
Teensy pins that do not yet have a function

SysTick
A type that represents the system timer, SYSTICK

Enums

Interrupt
Available interrupts for this device

interrupt
Available interrupts for this device

Constants

NVIC_PRIO_BITS
Number of priority bits implemented by the NVIC

Functions

configure_led
Configure the board's LED

Type Definitions

LED
The LED in its final configuration