Module teensy4_bsp::common

Common pinout

The Teensy 4.0 and 4.1 share many pins. This module provides the pins that are common across both boards. For pins that are unique to each board, and to acquire all of a board's pins, see the t40 and t41 modules.

Common pin table

This table is incomplete

We believe this table is accurate. But, there may be alternate functions that are not documented, and we're maintaining this table on a best-effort basis. Besides this table, there are two other ways to identify which pads support which peripheral:

study the i.MX RT 1060 Reference Manual. This is the authority on pad configuration.

study the trait implementations for the pad. Select a pin type alias, like P0, and click-through to its pad documentation (AD_B0_03). Notice the listing of imxrt-iomuxc trait implementations. This describes what kinds of functions the pin supports. The constraints may be enforced by the HAL's APIs.

We welcome documentation contributions!

Pin	Pad ID	Alt0	Alt1	Alt2	Alt3	Alt4	Alt5	Alt6	Alt7	Alt8	Alt9

0	AD_B0_03			UART6_RX		FlexPWM1_1_X

1	AD_B0_02			UART6_TX		FlexPWM1_0_X

2	EMC_04		FlexPWM4_2_A

3	EMC_05		FlexPWM4_2_B

4	EMC_06		FlexPWM2_0_A

5	EMC_08		FlexPWM2_1_A

6	B0_10			FlexPWM2_2_A

7	B1_01			UART4_RX				FlexPWM1_3_B

8	B1_00			UART4_TX				FlexPWM1_3_A

9	B0_11			FlexPWM2_2_B

10	B0_00				SPI4_PCS0

11	B0_02				SPI4_SDO

12	B0_01				SPI4_SDI

13	B0_03				SPI4_SCK		GPIO2_3 (LED)

14	AD_B1_02			UART2_TX

15	AD_B1_03			UART2_RX

16	AD_B1_07		I2C3_SCL	UART3_RX

17	AD_B1_06		I2C3_SDA	UART3_TX

18	AD_B1_01				I2C1_SDA

19	AD_B1_00			UART2_CTS	I2C1_SCL

20	AD_B1_10			UART8_TX

21	AD_B1_11			UART8_RX

22	AD_B1_08		FlexPWM4_0_A

23	AD_B1_09		FlexPWM4_1_A

24	AD_B0_12	I2C4_SCL		UART1_TX		FlexPWM1_2_X

25	AD_B0_13	I2C4_SDA		UART1_RX		FlexPWM1_3_X

26	AD_B1_14

27	AD_B1_15

28	EMC_32		FlexPWM3_1_B	UART7_RX

29	EMC_31		FlexPWM3_1_A	UART7_TX	SPI1_PCS1

30	EMC_37

31	EMC_36

32	B0_12

33	EMC_07		FlexPWM2_0_B

References:

Teensy Schematics

Type Definitions

P0
Pin 0 (common)

P1
Pin 1 (common)

P2
Pin 2 (common)

P3
Pin 3 (common)

P4
Pin 4 (common)

P5
Pin 5 (common)

P6
Pin 6 (common)

P7
Pin 7 (common)

P8
Pin 8 (common)

P9
Pin 9 (common)

P10
Pin 10 (common)

P11
Pin 11 (common)

P12
Pin 12 (common)

P13
Pin 13 (common)

P14
Pin 14 (common)

P15
Pin 15 (common)

P16
Pin 16 (common)

P17
Pin 17 (common)

P18
Pin 18 (common)

P19
Pin 19 (common)

P20
Pin 20 (common)

P21
Pin 21 (common)

P22
Pin 22 (common)

P23
Pin 23 (common)

P24
Pin 24 (common)

P25
Pin 25 (common)

P26
Pin 26 (common)

P27
Pin 27 (common)

P28
Pin 28 (common)

P29
Pin 29 (common)

P30
Pin 30 (common)

P31
Pin 31 (common)

P32
Pin 32 (common)

P33
Pin 33 (common)