Module teensy4_bsp::t40

Teensy 4.0 specific APIs

Use into_pins to constrain the processor pads into the pins available on the Teensy 4.0. If you cannot safely acquire all processor pads, use the unsafe Pins::new method to generate pins.

Pin	Pad ID	Alt0	Alt1	Alt2	Alt3	Alt4	Alt5	Alt6	Alt7	Alt8	Alt9

34	SD_B0_03		FlexPWM1_1_B			SPI1_SDI

35	SD_B0_02		FlexPWM1_1_A			SPI1_SDO

36	SD_B0_01		FlexPWM1_0_B	I2C3_SDA		SPI1_PCS0

37	SD_B0_00		FlexPWM1_0_A	I2C3_SCL		SPI1_SCK

38	SD_B0_05		FlexPWM1_2_B	UART8_RX

39	SD_B0_04		FlexPWM1_2_A	UART8_TX

Structs

Pins
Teensy 4.0 pins

Functions

into_pins
Constrain the processor pads to the Teensy 4.0 pins

Type Definitions

ErasedPins
Type-erased Teensy 4.0 pins

P0
Pin 0 (common)

P1
Pin 1 (common)

P2
Pin 2 (common)

P3
Pin 3 (common)

P4
Pin 4 (common)

P5
Pin 5 (common)

P6
Pin 6 (common)

P7
Pin 7 (common)

P8
Pin 8 (common)

P9
Pin 9 (common)

P10
Pin 10 (common)

P11
Pin 11 (common)

P12
Pin 12 (common)

P13
Pin 13 (common)

P14
Pin 14 (common)

P15
Pin 15 (common)

P16
Pin 16 (common)

P17
Pin 17 (common)

P18
Pin 18 (common)

P19
Pin 19 (common)

P20
Pin 20 (common)

P21
Pin 21 (common)

P22
Pin 22 (common)

P23
Pin 23 (common)

P24
Pin 24 (common)

P25
Pin 25 (common)

P26
Pin 26 (common)

P27
Pin 27 (common)

P28
Pin 28 (common)

P29
Pin 29 (common)

P30
Pin 30 (common)

P31
Pin 31 (common)

P32
Pin 32 (common)

P33
Pin 33 (common)

P34
Pin 34 (4.0)

P35
Pin 35 (4.0)

P36
Pin 36 (4.0)

P37
Pin 37 (4.0)

P38
Pin 38 (4.0)

P39
Pin 39 (4.0)