Crate teensy4_bsp

A Rust board support package (BSP) for the Teensy 4. Supports the Teensy 4.0 and 4.1 boards.

Peripherals are re-exported from the imxrt-rs hardware abstraction layer. See the HAL's documentation for more information on creating and using peripherals.

The BSP restricts the processor pads that are available, since the physical Teensy only has a few user-accessible pins. From these pins, you may construct peripherals and perform I/O. The two Teensy 4 boards support many of the same pins; see the common module for those similar pins. To construct Teensy 4.0 or 4.1 pins, see the corresponding into_pins function in each of the corresponding modules.

The BSP assumes some facilities of the processor, both which are required for the USB stack. Each are controllable through feature-flags. Each feature is on by default.

it registers the SysTick exception handler, and configures SYSTICK for a 1ms interrupt. Enabled with the "systick" feature, which is on by default.

it registers the USB_OTG1 interrupt, and uses the USB1 peripheral for logging. Enabled with the "usb-logging" feature, which is on by default. Depends on the "systick" feature.

If a user also registers a SysTick or USB_OTG1 handler, it may result in a duplicate definition error.

Re-exports

The BSP re-exports the following:

the teensy4-rt crate, as rt

the imxrt-hal crate, as hal

See the accompanying documentation of each crate for more information.

For simplicity, there may be other choice APIs from either crate that are re-exported in the BSP namespace.

Although it's not exported publicly, the BSP crate links in the teensy4-fcb crate, which provides a Firmware Configuration Block (FCB) specific to the Teensy 4. See the teensy4-fcb crate for details on FCBs.

Examples

Turn on a Teensy 4.0's LED:

#![no_std]
#![no_main]

extern crate panic_halt;

use bsp::rt::entry;
use cortex_m::asm::wfi;
use teensy4_bsp as bsp;

use embedded_hal::digital::v2::OutputPin;

#[entry]
fn main() -> ! {
    let peripherals = bsp::Peripherals::take().unwrap();
    let pins = bsp::t40::into_pins(peripherals.iomuxc);
    let mut led = bsp::configure_led(pins.p13);

    loop {
        led.set_high().unwrap();
        wfi();
    }
}

See the teensy4-examples crate for build-able, run-able examples. The examples utilize this BSP crate to blink LEDs, establish timers, and log data over USB.

Using RTIC

To develop Teensy 4 applications with the RTIC framework,

disables the teensy4-bsp's default features

enable the BSP's "rtic" feature

patch the cortex-m-rt crate with the Teensy 4's special runtime, available in the teensy4-bsp's repository.

All three steps can be summarized by this Cargo.toml snippet:

[dependencies.teensy4-bsp]
git = "https://github.com/mciantyre/teensy4-rs"
branch = "master"
default-features = false
features = ["rtic"]

[patch.crates-io.cortex-m-rt]
git = "https://github.com/mciantyre/teensy4-rs"
branch = "master"

You need to disable the BSP's default features, which disables the "systick" feature, to enable RTIC's SYSTICK handler. This means that you cannot use the USB logger when developing RTIC applications. Consider using the imxrt-uart-log crate for an alternate logging implementation.

You need to replace the typical cortex-m-rt runtime with our custom runtime. The patch above lets us replace RTIC's runtime crate with an API-compatible runtime.

Notice of alpha status

We've made some assumptions in this MVP BSP.

SYSTICK and delay implementation is very naive. Do not run for 49 continuous days, or risk a millisecond counter wrap-around.

Re-exports

pub use cortex_m_rt as rt;

pub use imxrt_hal as hal;

Modules

common
Common pinout

t40
Teensy 4.0 specific APIs

t41
Teensy 4.1 specific APIs

usb	usb-logging
Teensy 4 USB, taken from the original Teensy 4 C libraries.

Structs

Peripherals

SysTick	systick
A type that represents the system timer, SYSTICK

Enums

interrupt
Available interrupts for this device

Functions

configure_led
Configure the board's LED

Type Definitions

LED
The LED