Type Alias teensy4_bsp::board::LpuartBaud

pub type LpuartBaud = Baud;

Type for LPUART baud rates.

Use lpuart_baud() to easily compute a baud rate.

Aliased Type

struct LpuartBaud {
    pub osr: u32,
    pub sbr: u32,
    pub bothedge: bool,
}

Fields
osr: u32

Oversampling rate.

This should be set between 4 and 32. The driver clamps the osr value within this range.

sbr: u32

Baud rate modulo divisor.

The driver commits this value directly. A value of zero is allowed, but will disable baud rate generation in hardware. The max value is (2^13) - 1. The implementation limits the max value.

bothedge: bool

Both edge sampling.

Should be set when the oversampling rate is between 4 and 7. Optional for higher sampling rates. The driver will commit this value directly.