Module teensy4_bsp::usb

Teensy 4 USB, taken from the original Teensy 4 C libraries.

The USB stack provides a log implementation for logging over USB

This is Serial.println() in Rust. Use the macros of the log crate to write data over USB. Messages can be read back using screen or PuTTY.

Structs

LoggingConfig
Logging configuration

Reader
A type that can read USB serial messages from a host

USB
A handle that enables USB I/O

Writer
A type that can send data to a USB serial host