Function teensy4_bsp::usb::init

pub fn init(_: &SysTick, config: LoggingConfig) -> Result<Reader, Error>

This is supported on crate feature usb-logging only.

Initializes the USB stack. This prepares the logging back-end. Returns a Reader that can read USB serial messages.

To select the default logger behavior, specify Default::default() as the argument for config.

Before configuring the USB logger, you'll need to configure SysTick. Once you've configured SysTick, supply its reference here.

This may only be called once. If this is not called, we do not initialize the logger, and log messages will not be written to the USB host. Returns a SetLoggerError if the logging subsystem already has a logger.